Automatically calculate the adapter position of the item to expand
This also fixes a huge bug where the item-to-expand was earlier always expected to be present in the visual window. That is, its View was expected to be non-null. When the item View is null, the page is expanded from the top.

This will overall improve saving and restoring of state.