Fix: item change animations get canceled by ItemExpandAnimator
It's surprising that this bug flew under the radar for so long. ItemExpandAnimator registers a global layout listener on ExpandablePageLayout that was originally written to react to size changes and synchronize RecyclerView items with the page. As a side effect, it also used to override item add/remove/change/move animations. This commit modifies the implementation to ignore global layout callbacks unless the page's location or size changes.