Use CopyOnWriteArrayList for storing page callbacks
Reverse-for loops avoided object allocation while allowing callbacks to remove
themselves, but weren't foolproof. CopyOnWriteArrayList returns a new copy of
the underlying array on every iteration so it's safer. It's also more efficient
than using Kotlin's immutable list because it avoids blanket creating a new list
with extra capacity everytime a callback is added or removed.